rename lexer.c to analyzer.c
